pyproject.toml
[project]
name = "flask-buzz"
version = "4.0.0"
description = "Extra bindings on py-buzz specifically for flask apps"
authors = [
    {name = "Tucker Beck", email ="tucker.beck@gmail.com"},
]
readme = "README.md"
license-files = ["LICENSE.md"]
requires-python = ">=3.10"
dependencies = [
    "py-buzz~=7.0",
    "flask>=2.0,<4.0",
]

[project.urls]
homepage = "https://github.com/dusktreader/flask-buzz"
source = "https://github.com/dusktreader/flask-buzz"
changelog = "https://github.com/dusktreader/flask-buzz/blob/main/CHANGELOG.md"

[dependency-groups]
dev = [
    "basedpyright>=1.28.3",
    "ipython>=8.18.1",
    "mkdocs-material>=9.6.9",
    "mkdocstrings[python]>=0.29.0",
    "mypy>=1.15.0",
    "pyclean>=3.1.0",
    "pygments>=2.19.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-random-order>=1.1.1",
    "ruff>=0.11.0",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
minversion = "7.4.0"
addopts = [
    "--cov=src/flask_buzz",
    "--cov-report=term-missing",
    "--cov-fail-under=90",
    "--cov-report=xml:.coverage.xml",
    "--junitxml=.junit.xml",
    "--override-ini=junit_family=legacy",
]

[tool.ruff]
line-length = 120
src = ["src/flask_buzz", "tests"]

[tool.mypy]
pretty = true
files = ["src/flask_buzz", "tests"]

[[tool.mypy.overrides]]
module = ["flask_restx"]
ignore_missing_imports = true

[tool.basedpyright]
reportAny = false
reportExplicitAny = false
reportUnusedCallResult = false

[tool.hatch.build.targets.wheel]
packages = ["src/flask_buzz"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
